requirements-dev.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements-dev.in
#
astor==0.8.1
    # via flake8-simplify
attrs==21.4.0
    # via
    #   flake8-bugbear
    #   pytest
coverage[toml]==6.3.2
    # via pytest-cov
distlib==0.3.4
    # via virtualenv
filelock==3.6.0
    # via
    #   tox
    #   virtualenv
flake8==4.0.1
    # via
    #   -r requirements-dev.in
    #   flake8-bugbear
    #   flake8-isort
    #   flake8-simplify
flake8-bugbear==22.3.23
    # via -r requirements-dev.in
flake8-isort==4.1.1
    # via -r requirements-dev.in
flake8-simplify==0.19.2
    # via -r requirements-dev.in
iniconfig==1.1.1
    # via pytest
isort==5.10.1
    # via flake8-isort
mccabe==0.6.1
    # via flake8
mypy==0.942
    # via -r requirements-dev.in
mypy-extensions==0.4.3
    # via mypy
packaging==21.3
    # via
    #   pytest
    #   tox
platformdirs==2.5.1
    # via virtualenv
pluggy==1.0.0
    # via
    #   pytest
    #   tox
py==1.11.0
    # via
    #   pytest
    #   tox
pycodestyle==2.8.0
    # via flake8
pyflakes==2.4.0
    # via flake8
pyparsing==3.0.7
    # via packaging
pytest==7.1.1
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==3.0.0
    # via -r requirements-dev.in
pytest-django==4.5.2
    # via -r requirements-dev.in
six==1.16.0
    # via
    #   tox
    #   virtualenv
testfixtures==6.18.5
    # via flake8-isort
toml==0.10.2
    # via tox
tomli==2.0.1
    # via
    #   coverage
    #   mypy
    #   pytest
tox==3.24.5
    # via -r requirements-dev.in
typing-extensions==4.1.1
    # via mypy
virtualenv==20.14.0
    # via tox
